refactor: HeaderSection->Banner로 명칭 변경
HTML semantic element와 헷갈릴 가능성이 있음
